Apply patch ‘dwm-statuspadding-6.3.diff’
Fixed | Replaces magic numbers in statusbar with configurable variables.

horizpadbar for horizontal statusbar padding
vertpadbar for vertical statusbar padding

StatusText now has both left and right padding,
as well as the vertical padding that all of the statusbar shares.

Other than the addition of left padding to StatusText, appearance
of the statusbar is identical to pre-patch when using the defaults
in config.def.h